Added delete reserved slot

+++ reservation-service/reservation-application/src/main/java/com/azki/insurance/reservation/service/application/api/data/reservation/CreateReservationEdgeRequestDTO.java
@@ -16,5 +16,5 @@ public class CreateReservationEdgeRequestDTO extends BaseEdgeRequestDTO {
     @JsonProperty("slotIds")
     @NotEmpty(message = "At least one slot ID is required.")
     @Size(min = 1, max = 10, message = "The number of slot IDs must be between 1 and 10.")
-    private List<Integer> slotIds;
+    private List<Long> slotIds;
 }
