+++ reservation-service/reservation-data-access/src/main/java/com/azki/insurance/reservation/service/data/access/adapter/AvailableSlotsJpaAdapterImpl.java
@@ -17,7 +17,7 @@ import java.util.List;
 @Slf4j
 @Service
 public class AvailableSlotsJpaAdapterImpl extends BaseJpaAdapterImpl<AvailableSlotsDTO,
-                                                                     Integer,
+                                                                     Long,
                                                                      AvailableSlotsEntity,
                                                                      AvailableSlotsCriteriaDTO,
                                                                      AvailableSlotsJpaRepository>
