+++ reservation-service/reservation-data-access/src/main/java/com/azki/insurance/reservation/service/data/access/repository/AvailableSlotsJpaRepository.java
@@ -10,7 +10,7 @@ import java.util.Date;
 import java.util.List;
 
 @Repository
-public interface AvailableSlotsJpaRepository extends JpaRepository<AvailableSlotsEntity, Integer> {
+public interface AvailableSlotsJpaRepository extends JpaRepository<AvailableSlotsEntity, Long> {
 
     @Query(" SELECT S FROM AvailableSlotsEntity S WHERE  S.isReserved = FALSE AND S.startTime >= :startTime ORDER BY S.startTime ASC")
     List<AvailableSlotsEntity> findNearestAvailableSlot(Date startTime, Pageable pageable);
